Use CSS Flexbox for summary bar
Using Flexbox means the code does not have to compute percentages (which could
not fully sum up to 100% due to rounding errors) and can better catch how thin
columns render on narrow screens (by setting the flow-shrink property to 0 when
value is low).

The commit also automatically hides bars when the number of specs is 0.