Drop aom.idl file from report
See #10.

The AOM spec was converted into a placeholder with sub-specs per phase. Reffy
does not know how to handle specs with sub-specs, so failed to parse the IDL
that was moved and updated to the phase 1 sub-spec.

Now, the IDL in that sub-spec is actually a copy of the IDL defined in
WAI-ARIA 1.2 for now so should probably not appear in an aom.idl file for now
in the report (WAI-ARIA 1.2 should be added instead).