Split updates of ED and TR reports (#41)
The TR report is secondary, no need to push an update 4 times a day. The update
splits the update workflow into two workflows: one for the ED report and one for
the TR report.

The ED crawl runs 4 times per day and after a push on master as before.
The TR crawl runs 1 time per week on Monday.