[GitHub Actions] Drop useless config.json secret
The config.json file contained the API key that Reffy needed to interact with
the W3C API. Reffy no longer needs to send requests to the W3C API (these
requests are now done in w3c/browser-specs).

Defaults for other configuration settings in config.json should be good.

In short, the config.json file is useless, so let's not require it!