Set up publishing of @webref/idl with Lerna (#68)
Package gets prepared and published from the `packages/*` subfolder with Lerna.
Further packages are expected to be created.

Note this introduces a couple of line ending fixes for scripts to run on Windows:
- The .gitattributes file tells Git not to convert LF line endings of the bash
script. The script cannot run on typical Git configs on Windows otherwise.
- The `--binary` option tells `patch` to disable any CRLF to LF conversion
when applying the patches. That is also needed for patches to run smoothly on
Windows.

Fixes https://github.com/w3c/webref/issues/63.