Add logic to create @webref/css package
Code follows the same structure as for the `@webref/idl` package. Tests
make sure that all CSS fragments can be parsed with the Value Definition Syntax
parser of CSSTree.

Note that this is not yet the case, there are 10 cases with syntax errors that
require the creation of dedicated patches, not included in this commit.