Add jobs to semi-automate publication of @webref/idl package (#133)
This implements the following logic:

1. The `prepare-idl-release` job runs on Thursdays morning. It compares the
latest contents of the repo with the published `@webref/idl` package on NPM,
and creates a PR that bumps the version of the package in
`packages/idl/package.json`, and exposes the diff of the IDL.

2. The `release-idl` job runs whenever an update to `packages/idl/package.json`
gets pushed to the default branch. It publishes a new @webref/idl package to
npm and creates a new @webref/idl Git tag. The job is a no op if the version
number in package.json does not differ from the one published on npm.

As such, if all goes well, merging the PR created by the first job will trigger
the second job.