Stop running update-ed on pushes against default branch (#134)
The `update-ed` job ran on pushes against the default branch but that does not
make a lot of sense in practice. It also means that the job will run whenever
we bump the version of a package in preparation for a release, and the job may
thus clash with the release job.

The `update-ed` job can be triggered manually whenever needed.