Introduce improved release workflow (#211)
This updates implements the release workflow described in:
https://github.com/w3c/webref/issues/145#issuecomment-805668484

As opposed to the current release workflow, this workflow will release the
version that already appears in the `package.json` files, and not the bumped
version. Hence the patch bump on these files in this update.

The pre-release job takes care of creating a pre-release PR whenever needed. The
job updates the pre-release PR when the diff (or the version to release)
changes.

The created diff in the PR description is slightly more readable than the current one:
- New/Deleted files now appear at the beginning of the diff
- Individual diffs are separated by a newline

The job no longer runs on a fixed schedule, and rather runs whenever a push is
made that could create a diff with the current npm package. This may mean that
there will often be an open pre-release PR to review. If CSS and IDL change too
much, we could perhaps only assign reviewers to the pre-release PR only once per
week, instead of immediately when the PR gets created.

Reviewers get assigned to the PR once per week (on Thursdays) by a dedicated job.

The second job runs when a pre-release PR gets merged, and releases the package
to npm. The content of the package is taken at the main branch commit on which
the pre-release PR is based.