Fix workflows triggers
- Prevent pre-release workflows from running when pre-release PR gets merged
to avoid re-creating the PR before package gets released (#223):

There is no easy way to detect that a push was triggered by a PR merge, so hard
to both react on pushes *and* not react on a specific merge at the same time.
The workaround used here is not to react to changes to the file touched by the
pre-release PR. That is not perfect: if one bumps the version number manually
to release a major/minor version, the workflow will have to be triggered
manually as well to update the pre-release PR.

- Make sure pre-release workflows run after the "Update ED report" workflow has
pushed updates to the default branch (#219):

The workflows now use the "workflow_run" event. The "Update ED report" workflow
may not touch files that are in CSS/IDL packages, meaning that the workflows
will often run for nothing, but there's no easy way to avoid that.

- Make sure release workflow runs when pre-release PR gets merged (#222):

The `branches` constraint applies to the `base` branch, not to the `head`.
Check on the `head` branch now done in the `if` clause.