Pre-release workflows: rebase PR only when necessary (#242)
The pre-release PR was always rebased on the latest commit pushed to the
default branch, triggering push notifications each time the workflows were run.
This update only updates the pre-release PR when needed, meaning when the
version has changed or when the diff has changed.

Note that changes to `packages/xxx/package.json` that touch other parts than
the version won't be picked up. They need to be accompanied by a manual version
bump for the workflows to consider them.

Fixes #232.