Adjust workflows to release a @webref/elements package
This adjusts workflows to also handle and release a `@webref/elements` package,
as discussed in #51. Tests guarantee that interfaces that appear in the package
also exist in the `@webref/idl`.

Ideally, the IDL package should be added as peer dependency to the elements
package but that creates a chicken-and-egg situation as we will likely often
release packages together.

Patches are an overkill for now but could prove useful in the long run.

The update also splits CSS, elements, and IDL tests. That is a first step
towards addressing #237. It does not fully solve it though, as IDL patches are
still applied when the CSS package is released.