Add SVG 1.1 only interfaces to SVG patch
This is a best effort attempt to add missing SVG interfaces defined in SVG 1.1
and needed by SVG 1.1 elements. This is best effort in the sense that interfaces
had to be adjusted to pass IDL validation, and may not fully match the
interfaces shipped in browsers.

Some parts are clumsy too, such as the dependency on `CSSValue` (defined in
DOM Level 2) that is now also defined in the patch.

Fixes #278.