Add SVG 1.1 elements patch and adjust IDL patch (#281)
With the elements patch, only the SVG 1.1 elements that are still defined in
SVG 2 and those that are no longer in SVG 2 but are still implemented (altGlyph,
glyphRef, cursor and font) are preserved.

The IDL patch for SVG 1.1 has been adjusted accordingly to add related
interfaces defined in SVG 1.1. This is best effort in the sense that interfaces
had to be adjusted to pass IDL validation, and may not fully match the
interfaces shipped in browsers.

Fixes #278.