Release workflow: Install deps individually (#286)
To run the release script, the release workflow installed dependencies with
an `npm ci` call. Problem is that this triggers the prepare script as well,
applying CSS/IDL patches. If some patch is no longer valid because the
underlying spec got modified in the meantime, the workflow fails as a result.

That's a pity because the workflow does not depend on patches in practice. This
update installs deps that the workflow needs individually to avoid applying
patches.