Add small CLI tool to semi-automate creation of patches
This adds a quick and dirty create-patch tool to help with the creation of
patches. Tool can be called through one of:

```
node tools/create-patch.js
npm run create-patch
```

The tool takes for granted that last commit to current branch was to update one
CSS/elements/IDL extract file. It creates a patch file out of it, rolls back
that commit, copies the patch file to ed/csspatches or ed/idlpatches as needed,
and commits the file in a new branch.

Manual tasks that remain on top of preparing the initial commit are to review
the created patch, push it to GitHub, create the PR and get back to the
appropriate branch locally.