Get rid of "--delay" mocha CLI option (#364)
This is needed to fix #363. The `--delay` option requires the code to call the
`run` function only once and we were calling it multiple times. More
importantly, Mocha does not catch exceptions thrown at the root level and simply
reports nothing when that happens. These exceptions typically get thrown when
the IDL cannot be parsed, meaning that invalid data exceptions mostly went
unnoticed...

This update wraps all code that needs to run before tests in a `before`
function. Note the need to keep a "dummy" test at the root level, otherwise the
`before` function won't run (and wouldn't have anywhere to report failure).

This update slightly improves describe/it tests naming as well.