Bump version of CSS Tree in CSS package
This update bumps the version of CSS Tree referenced as peer dependency in the
CSS package. It also adjusts the parsing test exceptions (SVG Markers is no
longer referenced in browser-specs).

Major version of the CSS package itself gets bumped as well to account for the
major version bump in CSS Tree.