Drop IDL patch for Autoplay Policy Detection
... although note that an IDL update is pending to fix the fact that the
overloaded methods are not distinguishable (and cannot work as intended in any
case), see: https://github.com/w3c/autoplay/pull/22

Even though that problem is not caught by the Web IDL parser, we may want to
wait for the new IDL to land before we publish a new version of the npm package.
Or create another patch not to introduce Web IDL that we know is invalid.