Update tools and jobs to publish curated data (#468)
The goal of this update is to create a curated data view along with the raw data
view and the npm packages views (see #277).

Curation means applying patches to the raw data and re-generating the
`idlparsed`, `idlnames` and `idlnamesparsed` folders. The latter two will only
contain IDL names targeted at browsers, although note that actual spec filtering
remains a TODO at this stage (see corresponding TODO comments in
`prepare-curated.js` and `prepare-packages.js`).

To create the curated data view, this update introduces new tools:
- a `prepare-curated.js` tool that copies the raw data to the given folder,
applies patches (CSS, elements, IDL) when needed, re-generates the `idlparsed`
folder, re-generates the `idlnames` and `idlnamesparsed` folders and adjusts the
`index.json` and `idlnames.json` files accordingly.
- a `prepare-packages.js` tool (replaces the now gone `packages/prepare.js`)
that copies relevant curated data from the curated folder to the packages
folder.
- a `commit-curated.js` tool that updates the `curated` branch with the contents
of the given curated folder.

Goal is to have the `curated` branch be the one published as GitHub Pages.

The test logic was partially re-written to run the tests against the curated
data, and against both the curated data and the NPM packages data when tests
may yield different results.

A new `curate.yml` job publishes the curated data whenever the crawl data is
updated. The job also takes care or preparing package release PRs as needed,
replacing the previous prepare-xxx-release jobs.

The release workflow becomes:
1. Crawled data is updated (`update-ed.yml`)
2. Curated data and package data get generated (`curate.yml`)
3. Curated data and package data get tested (`curate.yml`)
4. The `curated` branch gets updated with the curated data (`curate.yml`)
5. Npm package pre-release PR gets created (`curate.yml`)
6. Someone reviews and merges the PR
7. New version of npm packages are released (`release-package.yml`)
8. A `Raw data for @webref/ttt@vx.y.z` tag gets added to the relevant commit on
the `main` branch.
9. A `@webref/ttt@vx.y.z` tag gets added to the relevant commit on the `curated`
branch.
10. The `@webref/ttt@latest` tag gets updated to point to the relevant commit on
the `curated` branch.

Note that, in order for a release to be created, curated data needs to have
changed.