Curation job: do not run on "packages/**" updates
The job should run when the `packages` folder gets updated but that will make
the job run immediately when we merge the release PR, and possibly before the
release job has a chance to complete.

This update makes the job ignore updates to the `packages` folder. This means
that the job will need to be triggered manually when such updates are made. Not
ideal but that was already the case before the curation job was introduced...