Use reffy fallback option to ignore erroneous data from errored crawls (#487)
Restricted to the ed crawl for the time being. To be extended to tr when it proves correct.

On top of extending to the tr crawl, the "Copy report" rsync command could be replaced
with a command that resets the target folder before it copies the files, so as to
automatically remove files that are no longer needed (c.f. #470). The clean job could then
be dropped altogether.