Drop extract files from specs that are no longer crawled (#490)
When a spec no longer appears in the list of specs crawled, we can be confident
that extract files that may have been created for it can be dropped from the
repo. This update adds a `clean-dropped-specs-files.js` script to take care of
that, and updates the "Update ED report" and "Update TR report" to run the
script before they commit the crawl results to the main branch.

This addresses #470.