Bump minor version of packages upon new/deleted files
When new/deleted files are detected in the CSS, elements, or IDL extracts, a
minor version bump should take place, which used to be done manually. This
update automates the process at the end of the crawl.

A manual bump is still required when extract files are deleted manually.

Fixes #341.