Allow static and regular operations with same name
See https://github.com/whatwg/webidl/issues/1097 for context.

This update adds a final check before reporting duplicated members to make
sure that the members are not a static and a regular operation with the
same identifier (now allowed in Web IDL).

This is needed to account for the recent IDL update in the Fetch API:
https://github.com/whatwg/fetch/commit/b3bfd0c877fd09da8c315b43cf1dd858f3ac1612