Adjust duplicate CSS handling rules and tests
The extraction of CSS value spaces based on `data-dfn-type` both removes a
handful of cases where we ended up with duplicate value spaces definitions in
CSS extracts, and produes a couple of additional ones for `<integer>`, `<url>`
and `<image>`. This update drops the exceptions-to-the-rules in the tests and
handles the additional duplicates generally by favoring definitions in CSS
Values and CSS Images.

The additional duplicates should rather be handled by Reffy, or at least Reffy
should report on the `data-dfn-for` attribute so that the location of the
global definition is, see:
https://github.com/w3c/reffy/issues/980