Add mechanism to patch event data
This provides a semi declarative approach to fix data about events which would be hard to fix in upstream specs.

Each 'patch' is described as a pattern to match, the number of expected matches (to detect when a patch effect may have changed) and the actual changes to make to the event data.

Currently implement changes on a subset of the specs that need so identified in https://github.com/w3c/webref/issues/630#issuecomment-1159028458