Add guarantees and related test for @webref/events package
This enforces the following guarantees:
- All events have a `type` attribute that match the name of the event
- All events have a `interface` attribute to describe the interface used by the
Event. The Web IDL interface exists and represents an actual interface
(i.e. not a mixin).
- All events have a `targets` attribute with a non-empty list of target
interfaces on which the event may fire. All Web IDL interfaces in the list
exist and represent an actual interface (i.e. not a mixin).
- The `bubbles` attribute is always set for events that can fire at interfaces
linked to a tree (DOM, IndexedDB, Web Bluetooth).
- The `bubbles` attribute is not set otherwise.

(Many tests currently fail. That's normal!)