Create getTreeInfo utility function
This creates a common `getTreeInfo` utility function that takes the name of an
interface as first parameter, an array of parsed known IDL interfaces (only
needed properties are `name` and `inheritance`) as second parameter, and that
returns a non-null object when the given interface is part of a tree hierarchy
on which events may bubble.

Returned object looks like:

```json
{
  "tree": "dom",
  "interface": "Element",
  "depth": 2
}
```

... with:
- `tree`: the shortname of the spec that defines the tree
- `interface`: the closest interface in the given interface's inheritance chain
that belongs to the tree hierarchy
- `depth`: the bubbling depth associated with the interface, where 0 means that
the interface is at the root of the hierarchy.

The new function is used in events tests.

The `amend-event-data` script still needs to be updated.