Adjust crawl/curation jobs for Reffy v8.0.0
This adjusts crawl and curation jobs to account for post-processing modules
introduced in Reffy v8.0.0.

The `csscomplete` post-processing module is run during the crawl because
we created CSS patches and de-duplication logic with completed CSS extracts in
mind. We could revisit that later on but that does not seem to be a priority.

The `idlparsed` and `idlnames` post-processing modules get run after IDL
patches have been applied to generate the `idlparsed`, `idlnames`, and
`idlnamesparsed` folder.

This update does not run the `events` post-processing module. To be included
in #650 along with the patch mechanism for events.

The curation logic leverages Reffy's new ability to copy crawl results to
another output folder. The whole logic could perhaps be further streamlined
to call Reffy's `crawlSpecs` function only once, e.g. by running patch
mechanisms as post-processing modules instead of treating them separately.
That also does not seem to be a priority...