Restrict amend-event-data script to patching events
The `events` post-processing module in Reffy now takes care of consolidating
the events into a single `events.json` file. The `amend-event-data.js` script
no longer needs to do it itself.

The update also drops the part that cleaned the `index.json` file as the
`prepare-curated.js` script already takes care of that.