Bump major version of CSS package to v5.0.0
Breaking change introduced by the switch to Reffy v9.0.0: at-rules are now
reported under a new `atrules` property (that itself is not a breaking change),
and descriptors are now to be found under the at-rule they apply to (so no more
`descriptors` properties at the root level).

Also note that all values that appear in the CSS package can now be parsed with
css-tree, meaning that there are no more exceptions.