Add schemas for all JSON extracts
This provides a first level of schema validation for curated data extracts,
see #657 for context.

Goal is to make it easier to detect and document (through a changelog, so also
useful for #704) situations where we change the structure of data extracts.

Schemas, notably those that deal with parsed IDL structures, could go deeper
into details.

Tests are run against the curated version of data. That is not necessary for
extracts that aren't actually curated (dfns, headings, ids, links, refs), just
more convenient not to have branching logic in the test code.