Fix CSS superseding relationship for css-values
Fixes #794.

css-values was incorrectly marked as being superseded by everything
else instead of actually superseding other specs. The list of specs being
superseded is now explicit.

Note this whole logic should soon more or less disappear with the next major
version of Reffy which supports `data-dfn-for` attributes and distinguishes
between CSS values that are namespaced from those that are not.