Fix name of extract file in CSS superseding logic
The code assumed that CSS extracts were always named after the spec's series'
shortname but that is not the case for specs that are not the current spec in
the series.

Note this bug was sitting around, patiently waiting for an occasion to reveal
itself. The recent fix on css-values superseding rules activated it because the
code now drops `<position>` from the css-backgrounds-4 export, and current
spec is css-backgrounds-3.