Rename "prepare" script to "curate" to prevent auto-runs
The `prepare` script is run when the repository is installed. On top of
creating issues when the repository is set as dependency in another
project for some reason, as described in:
https://github.com/w3c/webref/issues/789#issuecomment-1464094879

... this also seems wrong because:
1. There is no guarantee that the curation will run without errors. A patch
may no longer apply for instance.
2. Projects may want to depend on the raw data and may not need to run the
curation and package preparation logic at all.

This update replaces the "prepare" script with a "curate" one, explicitly
called in the jobs that need it.