Delegate Web IDL consistency tests to Strudy
With this update, all Web IDL consistency tests are now run directly by Strudy.
This allows to keep the test code simple and makes it possible to reuse the
test logic in other contexts.

Nothing was lost in the switch. Guarantees still hold in particular. Strudy
checks all IDL features that used to be checked in `consistency.js`, reporting
anomalies in a more digestable form (which should make it slightly easier to
find the root of a problem when one is found). Corresponding code in Strudy may
be found in https://github.com/w3c/strudy/blob/main/src/lib/study-webidl.js

Strudy reports a couple of Web IDL issues that the previous code did not detect,
namely the fact that the `Screen` interface needs to inherit from `EventTarget`,
for use in the Window Management API, and the fact that the `*Track` interfaces
need to be exposed on `DedicatedWorker`, for use in Media Source Extensions.
These issues have been acknowledged by the groups that develop the
specifications but the patch needs to be done against base interfaces that are
under the control of other groups, and it's not clear to me that these other
groups have acknowledged the issues yet. I did not feel like patching the IDL as
a result. The tests currently explicitly ignore these issues for now in the
interest of continuity.