Delegate Web IDL consistency tests to Strudy (#977)
With this update, all Web IDL consistency tests are now run directly by Strudy.
This allows to keep the test code simple and makes it possible to reuse the
test logic in other contexts.

Nothing was lost in the switch. Guarantees still hold in particular. Strudy
checks all IDL features that used to be checked in `consistency.js`, reporting
anomalies in a more digestable form (which should make it slightly easier to
find the root of a problem when one is found). Corresponding code in Strudy may
be found in https://github.com/w3c/strudy/blob/main/src/lib/study-webidl.js