Update captured-mouse-events patch
Former patch set the interface. The spec was fixed so interface no longer needs
to be explicitly specified. That said, the "does not bubble" wording is correct
but too complex for Reffy, who fails to parse "with its bubbles and cancelable
attributes set to false" and then thinks that "bubbles" means that the event
bubbles.