Only curate specs that have a good standing
This leverages the `standing` property in browser-specs to exclude specs that
don't have a good standing from data curation.

This makes it possible to add specs to browser-specs at an earlier level for
cross-referencing purpose (Specref and terms) without having to worry too much
about its impact on CSS, elements, events, and IDL definitions. It also makes it
possible to keep discontinued specs in browser-specs without having to worry
about extracts becoming obsolete, invalid, or conflictual.

This is intended to replace #712 with a different exclusion logic. In #712,
specs that were excluded from data curation were the ones that did not target
browsers, based on the `categories` property. This does not help with the main
source of CSS, events and IDL hiccups, which are more common in early API
proposals. Plus I still think that filtering specs based on their categories is
not the right approach.

This will remove the following curated extracts:
- CSS extract of CSS Conditional Values Module Level 1
- CSS extract of Non-element Selectors Module Level 1
- IDL extract of Direct Sockets API. If we want to keep the IDL, the right
mechanism would be to drop the "pending" standing, but the spec itself says that
it is an unofficial draft.
- IDL extract of Web Publications, which seems a good thing given that the spec
has been discontinued.
- IDL extract of Document Object Model (DOM) Level 2 Style, which we were
previously doing through a patch.

This does not add any mechanism to create exceptions to the rule. That is on
purpose. Let's be optimistic ;)

Note the plan to also make Reffy skip "discontinued" specs by default, in
https://github.com/w3c/reffy/pull/1341. With these two updates, the workflow
becomes:
1. Specs that are in good standing are crawled and curated
2. Specs that have a pending standing are crawled but not curated
3. Specs that have been discontinued are not crawled by default (but may be for
legacy/cross-referencing purpose) and not curated.

The rule for inclusion in NPM packages does not change: only specs targeted at
browsers are included.