Enable dependency caching in jobs
Hopefully, that will fix the transient Puppeteer installation errors that we
seem to get nowadays when jobs run.

Note this rewrites slightly the logic of the update jobs, but that should make
things slightly more straightforward (job could be further simplified to replace
the double checkout with a simple checkout, followed by a `git pull` after the
crawl).

The use of `npm ci` throughout the jobs means that we're now de facto pinning
down the version of Reffy that the crawler uses, as envisioned in #849. This
requires a manual action to bump the version in Webref when we release a new
version of Reffy but that's a cleaner approach.