Make css-borders-4 supersede css-backgrounds-3
For the context, see https://github.com/w3c/browser-specs/issues/1025

css-borders-4 extends some of the property declarations of css-backgrounds-3
(`border-xxx-radius`, `border-xxx-color`). This creates duplicate CSS property
definitions in Webref.

If we were to consider css-borders-4 as a delta spec, we would just do nothing
and leave de-duplication up to consumers. Here, we need to select one. If we
favor css-backgrounds-3 definitions, e.g., on the ground that they are more
stable, we would have to reverse that choice later on when that stops being the
case. Favoring css-borders-4 definitions seems more future-proof. Plus the
new definitions are backward-compatible with former ones.