Temporarily force crawl of discontinued RFCs
The status of some RFCs that used to be crawled by Reffy has recently (and
correctly) been switched to discontinued, see list in:
https://github.com/w3c/browser-specs/issues/327#issuecomment-1823278033

There are a bunch (*) of specs that still reference these discontinued specs.
This update makes Reffy continue to crawl the RFCs to give us more time to
update all specs, or at least to alert them that they need some update.

(*) I will prepare an explicit list of such specs.