Update README.md
Bitcoin is the world’s most traded cryptocurrency, and represents the largest piece of the crypto market pie. It was the first digital coin and as such, remains the most famous and widely-adopted cryptocurrency in the world. It's the original gangster in whose footsteps all other coins follow. The birth of Bitcoin was the genesis of an entirely new asset class, and a huge step away from traditional, centrally controlled money. Today, many advocates believe Bitcoin will facilitate the next stage for the global financial system, although this — of course — remains to be seen.