Set `cancelable` flag for "cancel" event (dialog and CloseWatcher) (#1237)
A recent HTML update changed the creation conditions for the "cancel" event
for `HTMLDialogElement` and `CloseWatcher`:
https://github.com/whatwg/html/pull/10291/files

The extraction code no longer detects the `cancelable` flag as a result.

In any case, the code that amended the event to split `HTMLElement` into
`HTMLInputElement` and `HTMLDialogElement` was incorrect: the `cancelable` flag
ended up on `HTMLInputElement` whereas it should have been set on the other two
target interfaces.

One interesting aspect of the HTML update is that it seems to make the event
cancelable depending on conditions. We cannot express that nuance with a
boolean flag.