Fix CSS test on base property definitions (#1273)
The test did not take into account the possibility that the base definition of
a CSS property could be in a delta spec. That's currently the case for the
`continue` property: defined in `css-overflow-4`, extended in `css-overflow-5`.

Not done here but the test could also be extended to make sure that the base
definition is defined in a delta spec that is logically *before* the spec that
extends the definition. The use of a separate array is meant to ease that
further update.

Closes #1272.