Bump css-tree from 2.3.1 to 3.1.0 (#1404)
Close #1378.

This also drops the CSS patch for css-values-5, no longer needed.

I still don't fully grasp why dependabot did not propose the bump but I suspect
that is due to the fact that we also have css-tree as a peer dependency for the
`@webref/css` package (which is also a dependency for testing purpose)