Add a validation guarantee to curated CDDL extracts (#1422)
* Add a validation guarantee to curated CDDL extracts

This leverages CDDL validation in Strudy to provide a validation guarantee to
curated CDDL extracts.

A `cddlpatches` folder, which will hopefully remain empty, gets added so that
we may temporarily patch CDDL extracts that become invalid.

* Adjust list of ignorable Web IDL anomalies

Bumping the version of Strudy brings new Web IDL anomaly types:
- `noEvent` needs events data that is not included in the data we pass to the
study function. It must be ignored.
- `urlType` is new and we probably don't want to block curation because of it,
at least for now (or we're going to have a lot of patching to do...)