Add "freeze patch" mechanism, add one for filter-effects-1 (#1505)
This creates a new way to curate the results of a crawl: a freeze patch
allows to freeze the results of a crawl to a specific commit ID in Webref. This
is meant for specs that are temporarily broken beyond repair.

First such patch is for the filter-effects-1 spec.

Each freeze patch is implemented as a JSON file named after the spec's
shortname and that contains the commit ID and a `pending` key that links to the
issue that tracks the problem in some GitHub repository. The `clean-patches`
script proposes to drop the patch when the issue gets closed.