Replace most CSS tests by analysis with Strudy
Strudy now integrates a set of analyses that were previously done in Webref.
It also extends slightly the duplication guarantees, which used to be
restricted to properties.

This update replaces most tests with running Strudy. What remains is tests that
assess that syntaxes can be parsed with css-tree. These validity tests would
better be done in Strudy too in practice, but that's not yet the case.

Note: This will report two new problems that need fixing (duplicate type
constructs in `css-color-5` and `css-color-hdr`). To be addressed with a patch.