Replace most CSS tests by analysis with Strudy (#1561)
* Replace most CSS tests by analysis with Strudy

Strudy now integrates a set of analyses that were previously done in Webref.
It also extends slightly the duplication guarantees, which used to be
restricted to properties.

This update replaces most tests with running Strudy. What remains is tests that
assess that syntaxes can be parsed with css-tree. These validity tests would
better be done in Strudy too in practice, but that's not yet the case.

* Improve duplication removal logic

This moves the check on delta specs after the core superseding logic so that
delta specs also get checked for actual duplicates with unrelated specs.