Add a few tests on consolidated CSS file (#1569)
This adds tests to guarantee that entries are unique in the consolidated CSS
file (they should be by construction, tests are meant to capture bugs that we
might miss if we update the consolidation logic in Reffy).

Consolidation amends syntaxes of some of the entries. Tests also make sure that
syntaxes can be parsed with CSSTree.