Create logic to publish two versions of `@webref/css`
Goal is to continue to publish the CSS extracts per spec as version 6 and to
publish the consolidated CSS as version 7 of the `@webref/css` package.

The versions are handled as two different packages in the code, but the
actual package name is the same (defined in respective `package.json` files).