Fix freeze patch branching logic
The code that applies freeze patches assumed that the caller would always want
to return to `main` afterwards. That assumption is only valid when curation
runs against `main`, not when it runs as part of a pull request, or locally
when preparing a pull request on a different branch.

The new logic gets back to the right position, either a branch name if there's
one, or a commit ID if contex is in "detached HEAD" mode.