Adjust `index()` in CSS package to also index descriptors
The consolidated CSS file is mostly flat except for descriptors that are nested
under at-rules. The descriptors were still using an array. For consistency, the
`index()` function now indexes them by name too.

README was also updated to drop the note on at-rules syntax expansion since we
no longer do that. It also notes that the list of descriptors is not guaranteed
to be exhaustive.